+++ delta_web/delta/frontend/src/components/DataSets/DatasetTable.js
@@ -15,7 +15,6 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { connect } from "react-redux";
-import { getCsvFiles } from "../../actions/datasets";
 import { Link } from "react-router-dom";
 
 // https://ui.dev/react-router-url-parameters
